Move all chessboard-related components to "chessboards" folder inside "features"
