+++ frontend/src/features/gameplay/chessboards/MultiplayerChessboard.tsx
@@ -37,12 +37,12 @@ import {
 	getRank,
 	getFile,
 	getSquareExists,
-} from "../../../utils/boardUtils";
+} from "../../../utils/boardUtils.ts";
 import {
 	fetchCurrentPosition,
 	fetchLegalMoves,
 	fetchMoveIsValid,
-} from "../../../utils/apiUtils";
+} from "../../../utils/apiUtils.ts";
 
 import {
 	whitePromotionRank,
@@ -56,7 +56,7 @@ import { getOppositeColor } from "../../../utils/gameLogic/general.ts";
 import useUsername from "../../../hooks/useUsername.ts";
 import usePieceAnimation from "../../../hooks/usePieceAnimation.ts";
 import { isObjEmpty } from "../../../utils/generalUtils.ts";
-import ChessboardGrid from "../../chessboard/ChessboardGrid.tsx";
+import ChessboardGrid from "../../../components/chessboard/ChessboardGrid.tsx";
 import {
 	EmptySquareRenderParams,
 	FilledSquareRenderParams,
